Pull `monkeypatch_env` into shared test configuration.
